Move playlist methods out of library.rs
